test: test_linear.py absorbs test_linear_filters.py
Linear's endpoint tests and its filter-compiler tests are one router (app/graphql/linear_*), so
one file with the filters as a section. The mutation-review rationale that justified the separate
file is kept in the module docstring — it explains why those assertions pin boundaries rather than
just 'it filters', which is the part a future reader needs.

test_endpoints.py 2270 -> 1897 lines. Same 860 passing.
