+++ tests/test_endpoints.py
@@ -1709,379 +1709,6 @@ def test_mock_openapi_spec_endpoint(client):
     assert client.get("/_mock/openapi/nope").status_code == 404
 
 
-# --- Linear (GraphQL) -------------------------------------------------------------
-# Linear is GraphQL-only, so there is no REST surface to crawl. What matters instead is that the
-# schema answers what real clients ask for: the LlamaIndex reader's exact field set, `@linear/sdk`'s
-# by-id relation roots, and Linear's own error/status split. (The TypeScript SDK itself is
-# exercised by the Node CI job — pytest cannot drive `@linear/sdk`.)
-
-def gql(client, query, headers, **variables):
-    body = {"query": query}
-    if variables:
-        body["variables"] = variables
-    return client.post("/linear/graphql", json=body, headers=headers)
-
-
-def linear_user_token(tokens_yaml, email):
-    return next(u["token"] for u in tokens_yaml["users"] if u["email"] == email)
-
-
-def lit(value) -> str:
-    """A GraphQL string literal. GraphQL only accepts DOUBLE quotes, so Python's %r (single
-    quotes) is a syntax error on the wire — json.dumps produces the right thing."""
-    return json.dumps(str(value))
-
-
-# The exact selection `llama-index-readers-linear` sends, and every field its `load_data()`
-# dereferences by subscript. A KeyError on any of them is the failure this guards.
-READER_QUERY = """
-query Team($id: String!) {
-  team(id: $id) {
-    issues {
-      nodes {
-        id title description createdAt updatedAt archivedAt autoArchivedAt autoClosedAt
-        branchName canceledAt completedAt dueDate estimate
-        creator { name } assignee { name } state { name } project { name }
-        labels { nodes { name } }
-      }
-    }
-  }
-}
-"""
-
-
-def test_linear_reader_field_set_all_resolves(client, admin_h):
-    r = gql(client, READER_QUERY, admin_h, id="ENG")
-    assert r.status_code == 200
-    assert "errors" not in r.json(), r.json().get("errors")
-    nodes = r.json()["data"]["team"]["issues"]["nodes"]
-    assert nodes
-    for issue in nodes:
-        # Present as KEYS even when null — the reader subscripts every one of them.
-        for field in ("id", "title", "description", "createdAt", "updatedAt", "archivedAt",
-                      "autoArchivedAt", "autoClosedAt", "branchName", "canceledAt",
-                      "completedAt", "dueDate", "estimate"):
-            assert field in issue, field
-        assert issue["labels"]["nodes"] is not None
-    # Key-presence alone is a TAUTOLOGY: graphql-core always emits a selected field as a key, so
-    # the loop above passes even if every value is served as a constant null. Pin the values that
-    # must be real, including a lifecycle timestamp that is genuinely populated.
-    done = next(i for i in nodes if i["title"] == "Continuous batching stalls after compaction")
-    assert done["completedAt"] == "2026-03-10T00:00:00Z"     # not None, not synthesized
-    assert done["createdAt"] == "2026-03-01T00:00:00Z"
-    assert done["canceledAt"] is None                        # Done, so it was never canceled
-    by_id = {i["title"]: i for i in nodes}
-    rl = by_id["Rate limiter drops bursts under 50ms"]
-    assert rl["creator"]["name"] and rl["assignee"]["name"] == "Bob Stone"
-    assert rl["state"]["name"] == "In Progress"
-    assert rl["project"]["name"] == "runtime-stability"
-    assert {label["name"] for label in rl["labels"]["nodes"]} == {"bug", "gateway"}
-    assert rl["estimate"] == 5
-    assert rl["dueDate"] == "2026-03-15"
-
-
-def test_linear_issue_by_uuid_and_by_identifier(client, admin_h):
-    by_key = gql(client, '{ issue(id: "ENG-101") { id identifier title } }', admin_h)
-    issue = by_key.json()["data"]["issue"]
-    assert issue["identifier"] == "ENG-101"
-    by_uuid = gql(client, "{ issue(id: %s) { identifier } }" % lit(issue["id"]), admin_h)
-    assert by_uuid.json()["data"]["issue"]["identifier"] == "ENG-101"
-
-
-def test_linear_missing_issue_is_a_field_error_not_a_400(client, admin_h):
-    """Linear declares `issue` non-null, so a miss nulls `data` and reports an error — but the
-    request itself was fine, so the status stays 200."""
-    r = gql(client, '{ issue(id: "NOPE-1") { identifier } }', admin_h)
-    assert r.status_code == 200
-    assert r.json()["data"] is None
-    assert "Entity not found" in r.json()["errors"][0]["message"]
-
-
-def test_linear_team_resolves_by_key_and_uuid(client, admin_h):
-    key = gql(client, '{ team(id: "ENG") { id key name } }', admin_h).json()["data"]["team"]
-    assert (key["key"], key["name"]) == ("ENG", "engineering")
-    assert gql(client, "{ team(id: %s) { key } }" % lit(key["id"]),
-               admin_h).json()["data"]["team"]["key"] == "ENG"
-
-
-def test_linear_team_issue_count_is_the_visible_count(client, admin_h, tokens_yaml):
-    """Asserted for BOTH an admin and a restricted caller: as admin alone the count's ACL branch
-    never runs, so the assertion would hold with scoping removed entirely."""
-    admin = {t["key"]: t["issueCount"] for t in
-             gql(client, "{ teams { nodes { key issueCount } } }",
-                 admin_h).json()["data"]["teams"]["nodes"]}
-    assert admin == {"ENG": 3, "DES": 1, "BLA": 1}
-    ava_h = {"Authorization": linear_user_token(tokens_yaml, "ava@acme.com")}
-    ava = {t["key"]: t["issueCount"] for t in
-           gql(client, "{ teams { nodes { key issueCount } } }",
-               ava_h).json()["data"]["teams"]["nodes"]}
-    # ava cannot see lin-secret or the blackops team at all.
-    assert ava == {"ENG": 2, "DES": 1}
-
-
-def test_linear_state_type_is_linears_category(client, admin_h):
-    r = gql(client, "{ issues { nodes { identifier state { name type } } } }", admin_h)
-    types = {n["identifier"]: n["state"]["type"] for n in r.json()["data"]["issues"]["nodes"]}
-    assert types == {"ENG-101": "started", "ENG-102": "completed", "DES-77": "started",
-                     "ENG-103": "backlog", "BLA-1": "triage"}
-
-
-def test_linear_priority_is_linears_numeric_scale(client, admin_h):
-    r = gql(client, "{ issues { nodes { identifier priority priorityLabel } } }", admin_h)
-    got = {n["identifier"]: (n["priority"], n["priorityLabel"])
-           for n in r.json()["data"]["issues"]["nodes"]}
-    # The corpus writes P0-P3; the API serves Linear's own 0-4 scale (1 = most urgent).
-    assert got["ENG-102"] == (1, "Urgent")
-    assert got["ENG-101"] == (2, "High")
-    assert got["DES-77"] == (3, "Medium")
-    assert got["ENG-103"] == (4, "Low")
-
-
-def test_linear_comments_connection_on_an_issue(client, admin_h):
-    r = gql(client, '{ issue(id: "ENG-101") { comments { nodes { body user { email } } } } }',
-            admin_h)
-    nodes = r.json()["data"]["issue"]["comments"]["nodes"]
-    assert [c["body"] for c in nodes] == ["Reproduced with a burst test.", "Fix is in review."]
-    assert nodes[0]["user"]["email"] == "bob@acme.com"
-
-
-def test_linear_by_id_relation_roots_answer(client, admin_h):
-    """`@linear/sdk` resolves relations lazily — `await issue.state` fires `workflowState(id:)`
-    rather than reading the value off the issue it already has. Without these roots every
-    relation accessor in the SDK fails."""
-    issue = gql(client, '{ issue(id: "ENG-101") { state { id } assignee { id } project { id } '
-                        'cycle { id } labels { nodes { id } } } }',
-                admin_h).json()["data"]["issue"]
-    assert gql(client, "{ workflowState(id: %s) { name team { key } } }" % lit(issue["state"]["id"]),
-               admin_h).json()["data"]["workflowState"] == {"name": "In Progress",
-                                                            "team": {"key": "ENG"}}
-    assert gql(client, "{ user(id: %s) { email } }" % lit(issue["assignee"]["id"]),
-               admin_h).json()["data"]["user"]["email"] == "bob@acme.com"
-    assert gql(client, "{ project(id: %s) { name } }" % lit(issue["project"]["id"]),
-               admin_h).json()["data"]["project"]["name"] == "runtime-stability"
-    assert gql(client, "{ cycle(id: %s) { name } }" % lit(issue["cycle"]["id"]),
-               admin_h).json()["data"]["cycle"]["name"] == "2025-W08"
-    label_id = issue["labels"]["nodes"][0]["id"]
-    assert gql(client, "{ issueLabel(id: %s) { name } }" % lit(label_id),
-               admin_h).json()["data"]["issueLabel"]["name"] in {"bug", "gateway"}
-
-
-def test_linear_workflow_states_are_per_team(client, admin_h):
-    """Two teams' identically-named states are different objects in Linear, so their ids differ.
-    The corpus has no shared state name, so assert the construction directly instead."""
-    from app import synth
-    assert synth.linear_state_id("Done", "engineering") != synth.linear_state_id("Done", "design")
-
-
-def test_linear_viewer_reports_the_authenticated_identity(client, tokens_yaml):
-    h = {"Authorization": linear_user_token(tokens_yaml, "ava@acme.com")}
-    me = gql(client, "{ viewer { email isMe } }", h).json()["data"]["viewer"]
-    assert me == {"email": "ava@acme.com", "isMe": True}
-
-
-def test_linear_content_round_trips_verbatim(client, admin_h, ro_conn):
-    """`Issue.description` is the doc's retrieval payload; it must come back byte-for-byte."""
-    stored = {r["identifier"]: r["content"]
-              for r in ro_conn.execute("SELECT identifier, content FROM linear_issues")}
-    r = gql(client, "{ issues(first: 100) { nodes { identifier description } } }", admin_h)
-    served = {n["identifier"]: n["description"] for n in r.json()["data"]["issues"]["nodes"]}
-    assert served == stored
-
-
-def test_linear_crawl_reaches_every_document(client, admin_h, ro_conn):
-    """The completeness assertion the REST crawls make, in Relay form: page with `first`/`after`
-    to exhaustion and land on exactly the stored row count."""
-    seen, cursor, guard = [], None, 0
-    while True:
-        guard += 1
-        assert guard < 50
-        after = (", after: %s" % lit(cursor)) if cursor else ""
-        page = gql(client, "{ issues(first: 2%s) { nodes { identifier } "
-                           "pageInfo { hasNextPage endCursor } } }" % after,
-                   admin_h).json()["data"]["issues"]
-        seen += [n["identifier"] for n in page["nodes"]]
-        if not page["pageInfo"]["hasNextPage"]:
-            break
-        cursor = page["pageInfo"]["endCursor"]
-    assert len(seen) == len(set(seen)) == db_count(ro_conn, "linear")
-
-
-def test_linear_introspection_reports_the_served_schema(client, admin_h):
-    r = gql(client, "{ __schema { queryType { name } mutationType { name } } "
-                    '__type(name: "Issue") { fields { name } } }', admin_h)
-    data = r.json()["data"]
-    assert data["__schema"]["queryType"]["name"] == "Query"
-    # Read-only mock: no Mutation root at all, rather than one advertising writes that fail.
-    assert data["__schema"]["mutationType"] is None
-    names = {f["name"] for f in data["__type"]["fields"]}
-    assert {"identifier", "branchName", "estimate", "dueDate", "state", "labels"} <= names
-
-
-def test_linear_malformed_document_is_a_400_with_a_graphql_envelope(client, admin_h):
-    r = gql(client, "{ issues(first: }", admin_h)
-    assert r.status_code == 400
-    body = r.json()
-    assert "detail" not in body and "data" not in body
-    assert "Syntax Error" in body["errors"][0]["message"]
-
-
-def test_linear_unauthenticated_is_401(client):
-    r = client.post("/linear/graphql", json={"query": "{ viewer { email } }"})
-    assert r.status_code == 401
-    assert r.json()["errors"][0]["message"] == "Authentication required"
-
-
-def test_linear_parent_resolves_and_is_acl_scoped(client, admin_h, tokens_yaml):
-    """`Issue.parent` is declared in the SDL and `@linear/sdk`'s fragment selects `parent { id }`.
-    The bench fills `parent_issue` on 46.7% of records, so it must resolve — and it must resolve
-    through the ACL, or it becomes another way to confirm a hidden issue exists."""
-    # lin-batch (ENG-102) is parented to lin-secret (ENG-103), which only hana can read.
-    q = '{ issue(id: "ENG-102") { identifier parent { identifier title } } }'
-    as_hana = gql(client, q, {"Authorization": linear_user_token(tokens_yaml, "hana@acme.com")})
-    assert as_hana.json()["data"]["issue"]["parent"]["identifier"] == "ENG-103"
-    as_ava = gql(client, q, {"Authorization": linear_user_token(tokens_yaml, "ava@acme.com")})
-    assert as_ava.json()["data"]["issue"]["parent"] is None    # hidden parent -> null, not a leak
-    # admin sees it, confirming the null above is the ACL and not a broken lookup
-    assert gql(client, q, admin_h).json()["data"]["issue"]["parent"]["identifier"] == "ENG-103"
-
-
-def test_linear_issue_without_a_parent_is_null(client, admin_h):
-    assert gql(client, '{ issue(id: "ENG-101") { parent { identifier } } }',
-               admin_h).json()["data"]["issue"]["parent"] is None
-
-
-def test_linear_default_ordering_is_by_creation_not_insertion(client, admin_h):
-    """Linear's docs: "By default results are ordered by createdAt field." An absent `orderBy`
-    previously fell through to raw insertion order, so `issues(first: n)` returned an arbitrary n
-    rather than the first n by creation."""
-    q = "{ issues(first: 50%s) { nodes { identifier createdAt } } }"
-    default = [n["createdAt"] for n in gql(client, q % "", admin_h).json()["data"]["issues"]["nodes"]]
-    explicit = [n["createdAt"] for n in
-                gql(client, q % ", orderBy: createdAt", admin_h).json()["data"]["issues"]["nodes"]]
-    assert default == explicit
-    assert default == sorted(default), "default ordering must be by creation, ascending"
-
-
-def test_linear_sort_input_overrides_the_default_ordering(client, admin_h):
-    """`orderBy` carries no direction in Linear, so `sort:` is how a client asks for the other
-    one — which means it has to actually win over the default."""
-    q = ('{ issues(first: 50, sort: [{createdAt: {order: Descending}}]) '
-         "{ nodes { createdAt } } }")
-    got = [n["createdAt"] for n in gql(client, q, admin_h).json()["data"]["issues"]["nodes"]]
-    assert got == sorted(got, reverse=True)
-
-
-# --- Linear relations / children / attachments / releases (#25) -----------------------
-
-def test_linear_children_is_the_exact_inverse_of_parent(client, admin_h):
-    """Linear DEFINES `children` as the inverse of `parent`, so the two must never disagree. They
-    are both read off the `parent_doc_id` resolved at import rather than joined on `identifier`,
-    because bench keys repeat — a join would attach one issue's children to every issue sharing
-    its key."""
-    kids = gql(client, '{ issue(id: "ENG-103") { children { nodes { identifier } } } }',
-               admin_h).json()["data"]["issue"]["children"]["nodes"]
-    assert [k["identifier"] for k in kids] == ["ENG-102"]
-    back = gql(client, '{ issue(id: "ENG-102") { parent { identifier } } }',
-               admin_h).json()["data"]["issue"]["parent"]
-    assert back["identifier"] == "ENG-103"
-
-
-def test_linear_children_is_acl_scoped(client, tokens_yaml):
-    """ENG-103 is restricted to hana, so ava cannot even reach it to ask for its children — and
-    the children list must never become a way to observe an issue she is denied."""
-    ava = {"Authorization": linear_user_token(tokens_yaml, "ava@acme.com")}
-    denied = gql(client, '{ issue(id: "ENG-103") { children { nodes { identifier } } } }', ava)
-    assert "Entity not found" in denied.json()["errors"][0]["message"]
-
-
-def test_linear_relations_and_their_inverse(client, admin_h):
-    rels = gql(client, '{ issue(id: "ENG-102") { relations { nodes { type relatedIssue '
-                       "{ identifier } } } } }", admin_h).json()["data"]["issue"]["relations"]["nodes"]
-    assert sorted((r["type"], r["relatedIssue"]["identifier"]) for r in rels) == \
-        [("blocks", "ENG-101"), ("related", "ENG-103")]
-    # the same row read from the other end
-    inv = gql(client, '{ issue(id: "ENG-101") { inverseRelations { nodes { type issue '
-                      "{ identifier } } } } }", admin_h).json()["data"]["issue"]
-    assert [(r["type"], r["issue"]["identifier"]) for r in inv["inverseRelations"]["nodes"]] == \
-        [("blocks", "ENG-102")]
-
-
-def test_linear_relation_to_a_hidden_issue_is_omitted(client, tokens_yaml, admin_h):
-    """A relation is scoped on the FAR end: surfacing one whose counterpart the caller cannot read
-    would disclose that issue's existence — the leak class the by-id roots were fixed for."""
-    ava = {"Authorization": linear_user_token(tokens_yaml, "ava@acme.com")}
-    q = '{ issue(id: "ENG-102") { relations { nodes { relatedIssue { identifier } } } } }'
-    seen = [r["relatedIssue"]["identifier"]
-            for r in gql(client, q, ava).json()["data"]["issue"]["relations"]["nodes"]]
-    assert seen == ["ENG-101"], "the relation to the restricted ENG-103 must be omitted"
-    # admin sees both, proving the omission is the ACL and not a broken join
-    assert len(gql(client, q, admin_h).json()["data"]["issue"]["relations"]["nodes"]) == 2
-
-
-def test_linear_attachments_from_both_bench_shapes(client, admin_h):
-    """`Attachment.title` is non-null in Linear, so a bare URL needs a derived title rather than
-    an empty string."""
-    nodes = gql(client, '{ issue(id: "ENG-102") { attachments { nodes { title url } } } }',
-                admin_h).json()["data"]["issue"]["attachments"]["nodes"]
-    got = {n["title"]: n["url"] for n in nodes}
-    assert got["Design doc"] == "https://conf.acme.test/design/batching"   # explicit title
-    assert got["artifacts.zip"] == "https://ci.acme.test/builds/4821/artifacts.zip"  # derived
-
-
-def test_linear_attachments_url_argument_and_filter(client, admin_h):
-    one = gql(client, '{ issue(id: "ENG-102") { attachments(url: "https://conf.acme.test/design/'
-                      'batching") { nodes { title } } } }', admin_h)
-    assert [n["title"] for n in one.json()["data"]["issue"]["attachments"]["nodes"]] == ["Design doc"]
-    none = gql(client, '{ issue(id: "ENG-102") { attachments(filter: {title: {eq: "nope"}}) '
-                       "{ nodes { title } } } }", admin_h)
-    assert none.json()["data"]["issue"]["attachments"]["nodes"] == []
-
-
-def test_linear_releases_and_the_by_id_root(client, admin_h):
-    nodes = gql(client, '{ issue(id: "ENG-102") { releases { nodes { id name slugId } } } }',
-                admin_h).json()["data"]["issue"]["releases"]["nodes"]
-    assert [n["name"] for n in nodes] == ["runtime-1.19"]
-    assert gql(client, '{ release(id: %s) { name } }' % lit(nodes[0]["id"]),
-               admin_h).json()["data"]["release"]["name"] == "runtime-1.19"
-
-
-def test_linear_release_by_id_is_acl_scoped(client, tokens_yaml):
-    """The release only appears on ENG-102, which ava CAN read — so she resolves it. Asserted to
-    pin that the scoping is on visibility, not a blanket denial."""
-    from app import synth
-    ava = {"Authorization": linear_user_token(tokens_yaml, "ava@acme.com")}
-    got = gql(client, '{ release(id: %s) { name } }' % lit(synth.linear_release_id("runtime-1.19")),
-              ava)
-    assert got.json()["data"]["release"]["name"] == "runtime-1.19"
-    absent = gql(client, '{ release(id: %s) { name } }' % lit(synth.linear_release_id("nope-9")), ava)
-    assert "Entity not found" in absent.json()["errors"][0]["message"]
-
-
-def test_linear_issue_with_no_relations_returns_empty_connections(client, admin_h):
-    r = gql(client, '{ issue(id: "DES-77") { relations { nodes { id } } children { nodes { id } } '
-                    "attachments { nodes { id } } releases { nodes { id } } } }",
-            admin_h).json()["data"]["issue"]
-    assert all(r[k]["nodes"] == [] for k in ("relations", "children", "attachments", "releases"))
-
-
-def test_linear_parent_and_children_read_the_same_column(client, admin_h, ro_conn):
-    """Both directions must consult the resolved `parent_doc_id`, not two independent lookups
-    that happen to agree — that is the whole reason the key is resolved once at import.
-
-    Also a performance contract: `@linear/sdk`'s Issue fragment selects `parent { id }` on every
-    node, so resolving it by identifier cost ~45ms on a 50-issue page."""
-    # `ro_conn` is the SAMPLE db; a fresh get_settings() would follow whatever MOCK_DATA_DIR
-    # another module last set, which is why this reads the fixture instead.
-    row = ro_conn.execute("SELECT doc_id, parent_doc_id, parent_key FROM linear_issues "
-                          "WHERE doc_id = 'lin-batch'").fetchone()
-    # the import pass resolved the KEY into a doc_id
-    assert row["parent_key"] == "ENG-103"
-    assert row["parent_doc_id"] == "lin-secret"
-    served = gql(client, '{ issue(id: "ENG-102") { parent { identifier } } }',
-                 admin_h).json()["data"]["issue"]["parent"]
-    assert served["identifier"] == "ENG-103"
-
-
 # --- fireflies: POST /fireflies/graphql -----------------------------------------
 # Fireflies has no SDK and no LlamaIndex reader; the vendor's own quickstart is a raw HTTP POST,
 # so this IS the client story rather than a fallback for one.
